fix: add jsdoc and rename file
